requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pyperclip>=1.8.2
